feat: refactor fc section progress
