+++ frontend/src/assets/styles/styles.scss
@@ -17,10 +17,6 @@ body {
 }
 
 /* ----------------- form input styles --------------------------- */
-.form-field {
-    margin-bottom: 1.5rem;
-}
-
 .input-with-verify-button {
     padding: 0;
     width: 100%;
